requirements.txt
Flask>=3.0
dnspython>=2.6
aiosmtpd>=1.4
pytest>=7.0
pytest-asyncio>=0.21
